Enables in-place video rendering option
Introduces a preprocessor definition `WCV_INPLACE` to allow for in-place video rendering within a specified image window.

This change provides flexibility in how video frames are displayed, allowing for direct rendering into an existing UI element instead of relying on a separate window. Also fixes a crash when no decoder is set.